+++ deploy/kubernetes/keytransparency-deployment.yml.tmpl
@@ -90,7 +90,7 @@ spec:
     spec:
       containers:
       - name: kt-signer
-        image: us.gcr.io/key-transparency/keytransparency-signer
+        image: us.gcr.io/key-transparency/keytransparency-sequencer
         imagePullPolicy: Always
         ports:
           - containerPort: 8080
